oh-tab-11z: split useHostMessages
